scripts/requirements.txt
affine==2.4.0
attrs==25.4.0
certifi==2025.11.12
cftime==1.6.5
click==8.3.1
cligj==0.7.2
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
dask==2025.12.0
dotenv==0.9.9
fonttools==4.61.1
fsspec==2025.12.0
kiwisolver==1.4.9
locket==1.0.0
matplotlib==3.10.8
netCDF4==1.7.3
numpy==2.4.0
packaging==25.0
pandas==2.3.3
partd==1.4.2
pillow==12.0.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rasterio==1.5.0
six==1.17.0
toolz==1.1.0
tzdata==2025.3
xarray==2025.12.0
